add db module, add config stores for saving wallet and peer cache
